workflow_files: move install functionality into its own module
